+++ Nos/Service/CurrentUser.swift
@@ -28,28 +28,40 @@ enum CurrentUser {
         return nil
     }
     
+    static var context: NSManagedObjectContext?
+    
     static var relayService: RelayService? {
         didSet {
             if let pubKey = publicKey {
-                let filter = Filter(authorKeys: [pubKey], kinds: [.metaData, .contactList], limit: 2)
-                relayService?.requestEventsFromAll(filter: filter)
+                // Load contact list into memory from Core Data
+                if let context = context,
+                    let author = author(in: context),
+                    let follows = author.follows as? Set<Follow> {
+                    CurrentUser.follows = follows
+                }
+                
+                // Refresh contact list and metadata from the relays
+                let metadataFilter = Filter(authorKeys: [pubKey], kinds: [.metaData], limit: 1)
+                relayService?.requestEventsFromAll(filter: metadataFilter)
+                let contactListFilter = Filter(authorKeys: [pubKey], kinds: [.contactList], limit: 1)
+                relayService?.requestEventsFromAll(filter: contactListFilter)
             }
         }
     }
     
-    static var follows: [Follow]?
+    static var follows: Set<Follow>?
     
     static func isFollowing(key: String) -> Bool {
         guard let following = follows else {
             return false
         }
         
-        let followKeys = following.map({ $0.identifier })
+        let followKeys = following.compactMap({ $0.destination?.hexadecimalPublicKey })
         return followKeys.contains(key)
     }
     
-    static func refresh() {
-        var authors = follows?.map { $0.identifier! } ?? []
+    static func refreshHomeFeed() {
+        var authors = follows?.compactMap { $0.destination?.hexadecimalPublicKey } ?? []
         if let pubKey = publicKey {
             authors.append(pubKey)
         }
@@ -71,23 +83,25 @@ enum CurrentUser {
             relayString += "{\"\(relay)\":{\"write\":true,\"read\":true}"
         }
         
-        // swiftlint:disable line_length
         let time = Int64(Date.now.timeIntervalSince1970)
         let kind = EventKind.contactList.rawValue
-        let jsonEvent = JSONEvent(id: "0", pubKey: pubKey, createdAt: time, kind: kind, tags: tags, content: relayString, signature: "")
-        // swiftlint:enable line_length
-        
-        let event = Event(context: context, jsonEvent: jsonEvent)
-        event.identifier = try? event.calculateIdentifier()
+        var jsonEvent = JSONEvent(pubKey: pubKey, createdAt: time, kind: kind, tags: tags, content: relayString)
         
         if let privateKey = privateKey, let pair = KeyPair(privateKeyHex: privateKey) {
-            try? event.sign(withKey: pair)
-            relayService?.sendEventToAll(event: event)
+            do {
+                try jsonEvent.sign(withKey: pair)
+                let event = try EventProcessor.parse(jsonEvent: jsonEvent, in: PersistenceController.shared)
+                relayService?.sendEventToAll(event: event)
+            } catch {
+                print("failed to update Follows \(error.localizedDescription)")
+            }
         }
         
         // Refresh contact list and meta data
-        let filter = Filter(authorKeys: [pubKey, followKey], kinds: [.contactList, .metaData], limit: 4)
-        relayService?.requestEventsFromAll(filter: filter)
+        let metadataFilter = Filter(authorKeys: [pubKey], kinds: [.metaData], limit: 1)
+        relayService?.requestEventsFromAll(filter: metadataFilter)
+        let contactListFilter = Filter(authorKeys: [followKey], kinds: [.contactList], limit: 1)
+        relayService?.requestEventsFromAll(filter: contactListFilter)
     }
     
     /// Follow by public hex key
@@ -99,40 +113,53 @@ enum CurrentUser {
 
         print("Following \(key)")
 
-        var followKeys = follows?.map { $0.identifier! } ?? []
+        var followKeys = follows?.compactMap { $0.destination?.hexadecimalPublicKey } ?? []
         followKeys.append(key)
         let tags = followKeys.map { ["p", $0] }
         
         updateFollows(pubKey: pubKey, followKey: key, tags: tags, context: context)
         
         // Refresh everyone's meta data and contact list
-        refresh()
+        refreshHomeFeed()
     }
     
     /// Unfollow by public hex key
-    static func unfollow(key: String, context: NSManagedObjectContext) {
+    static func unfollow(key unfollowedKey: String, context: NSManagedObjectContext) {
         guard let pubKey = publicKey else {
             print("Error: No pubkey for current user")
             return
         }
 
-        print("Unfollowing \(key)")
+        print("Unfollowing \(unfollowedKey)")
         
-        let followKeys = follows?.filter { $0.identifier! != key } ?? []
-        let followStrings = followKeys.map { $0.identifier! }
-        let tags = followStrings.map { ["p", $0] }
+        let stillFollowingKeys = (follows ?? [])
+            .compactMap { $0.destination?.hexadecimalPublicKey }
+            .filter { $0 != unfollowedKey }
+        let tags = stillFollowingKeys.map { ["p", $0] }
 
-        updateFollows(pubKey: pubKey, followKey: key, tags: tags, context: context)
+        updateFollows(pubKey: pubKey, followKey: unfollowedKey, tags: tags, context: context)
 
         // Delete cached texts from this person
-        if let author = try? Author.find(by: key, context: context) {
+        if let author = try? Author.find(by: unfollowedKey, context: context) {
             let deleteRequest = Event.deleteAllPosts(by: author)
             
             do {
                 try context.execute(deleteRequest)
             } catch let error as NSError {
-                print("Failed to delete texts from \(key). Error: \(error.description)")
+                print("Failed to delete texts from \(unfollowedKey). Error: \(error.description)")
             }
+        }
+    }
+    
+    static func author(in context: NSManagedObjectContext) -> Author? {
+        if let publicKey = Self.publicKey {
+            let author = try? Author.findOrCreate(by: publicKey, context: context)
+            if context.hasChanges {
+                try? context.save()
+            }
+            return author
+        } else {
+            return nil
         }
     }
 }